[CHORE] Splitted different webpack configs.
