+++ .npmrc
@@ -1,1 +1,3 @@
 save-exact=true
+loglevel=verbose
+color=always
