--- .travis.yml
@@ -1,11 +1,0 @@
-language: node_js
-cache: yarn
-node_js:
-  - "7"
-  - "6"
-  - "5"
-  - "4"
-script: yarn run test:CI
-after_success:
-  # Send coverage data to coveralls.
-  - yarn run test:cover:CI
