--- appveyor.yml
@@ -1,39 +1,0 @@
-# Test against this version of Node.js
-environment:
-  matrix:
-  # node.js
-  - nodejs_version: "7"
-  - nodejs_version: "6"
-  - nodejs_version: "5"
-  - nodejs_version: "4"
-
-# Install scripts. (runs after repo cloning)
-install:
-  # Get the latest stable version of Node.js or io.js
-  - ps: Install-Product node $env:nodejs_version
-  # install modules
-  - yarn
-
-cache:
-  - "%LOCALAPPDATA%\\Yarn"
-
-# Post-install test scripts.
-test_script:
-  # Output useful info for debugging.
-  - node --version
-  - npm --version
-  # run tests
-  - npm run test:CI
-
-# Don't actually build.
-build: off
-
-notifications:
-- provider: Email
-  to:
-  - housecor@gmail.com
-  subject: 'Build failed: react-slingshot'
-  message: The continuous integration build failed. See https://ci.appveyor.com/project/CoryHouse/react-slingshot/ for details.
-  on_build_success: false
-  on_build_failure: true
-  on_build_status_changed: false
