+++ config/webpack.config.prod.js
@@ -1,0 +1,54 @@
+import webpack from 'webpack';
+import merge from 'webpack-merge';
+import ExtractTextPlugin from 'extract-text-webpack-plugin';
+import WebpackMd5Hash from 'webpack-md5-hash';
+import HtmlWebpackPlugin from 'html-webpack-plugin';
+import path from 'path';
+
+import sharedConfig from './webpack.config';
+
+const GLOBALS = {
+  'process.env.NODE_ENV': JSON.stringify('production'),
+  __DEV__: false
+};
+
+export default merge(sharedConfig, {
+  // Emit a full SourceMap file.
+  devtool: 'source-map',
+  entry: path.resolve(__dirname, '..', 'src/index'),
+  output: {
+    path: path.resolve(__dirname, '..', 'dist'),
+    publicPath: '/',
+    filename: '[name].[chunkhash].js'
+  },
+  plugins: [
+    // Hash the files using MD5 so that their names change when the content changes.
+    new WebpackMd5Hash(),
+    // Tells React to build in prod mode. https://facebook.github.io/react/downloads.html
+    new webpack.DefinePlugin(GLOBALS),
+    // Generate an external css file with a hash in the filename
+    new ExtractTextPlugin('[name].[contenthash].css'),
+    // Generate HTML file that contains references to generated bundles. See here for how this works: https://github.com/ampedandwired/html-webpack-plugin#basic-usage
+    new HtmlWebpackPlugin({
+      template: path.resolve(__dirname, '..', 'src/index.ejs'),
+      favicon: path.resolve(__dirname, '..', 'src/favicon.ico'),
+      minify: {
+        removeComments: true,
+        collapseWhitespace: true,
+        removeRedundantAttributes: true,
+        useShortDoctype: true,
+        removeEmptyAttributes: true,
+        removeStyleLinkTypeAttributes: true,
+        keepClosingSlash: true,
+        minifyJS: true,
+        minifyCSS: true,
+        minifyURLs: true
+      },
+      inject: true
+    }),
+    // Minify JS
+    new webpack.optimize.UglifyJsPlugin({
+      sourceMap: true
+    })
+  ]
+});
