+++ config/webpackLoaders/babel.js
@@ -1,0 +1,5 @@
+module.exports = {
+  test: /\.jsx?$/,
+  exclude: /node_modules/,
+  use: ['babel-loader']
+};
