+++ config/webpackLoaders/file.js
@@ -1,0 +1,4 @@
+module.exports = {
+  test: /\.eot(\?v=\d+.\d+.\d+)?$/,
+  use: ['file-loader']
+};
