+++ config/webpackLoaders/image.js
@@ -1,0 +1,11 @@
+module.exports = {
+  test: /\.(jpe?g|png|gif|ico)$/i,
+  use: [
+    {
+      loader: 'file-loader',
+      options: {
+        name: '[name].[ext]'
+      }
+    }
+  ]
+};
