+++ package.json
@@ -8,14 +8,13 @@
     "remove-demo": "babel-node tools/removeDemo.js",
     "start-message": "babel-node tools/startMessage.js",
     "prestart": "npm run start-message",
-    "start": "concurrently -k -r -s first \"npm run test:watch\" \"npm run open:src\" \"npm run lint:watch\"",
+    "start": "concurrently -k -r -s first \"npm run test:watch\" \"npm run open:src\"",
     "open:src": "babel-node tools/srcServer.js",
     "open:dist": "babel-node tools/distServer.js",
     "lint": "eslint src/ tools --color",
-    "lint:watch": "npm run lint -- --watch",
+    "lint:watch": "npm run lint --watch",
     "clean-dist": "npm run remove-dist && mkdir dist",
     "remove-dist": "rimraf ./dist",
-    "prebuild": "npm run lint && npm run test && npm run clean-dist",
     "build": "babel-node tools/build.js && npm run open:dist",
     "test": "jest --config config/jest.json",
     "test:CI": "babel-node tools/testCi.js",
@@ -35,7 +34,8 @@
     "react-router-dom": "4.2.2",
     "react-router-redux": "5.0.0-alpha.6",
     "redux": "3.7.2",
-    "redux-thunk": "2.2.0"
+    "redux-thunk": "2.2.0",
+    "webpack-merge": "4.1.1"
   },
   "devDependencies": {
     "@nimbl3/eslint-config-nimbl3": "2.0.1",
@@ -67,6 +67,7 @@
     "eslint-watch": "3.1.3",
     "extract-text-webpack-plugin": "3.0.2",
     "file-loader": "1.1.6",
+    "glob": "7.1.2",
     "history": "4.7.2",
     "html-webpack-plugin": "2.30.1",
     "identity-obj-proxy": "3.0.0",
